refactor(docs): migrate Base and Bitable endpoint semantics (#439)
