+++ crates/openlark-docs/src/base/base/v2/app/role/delete.rs
@@ -82,15 +82,17 @@ mod tests {
     use super::*;
     use serde_json::json;
     use wiremock::MockServer;
-    use wiremock::matchers::{method, path};
+    use wiremock::matchers::{header, method, path};
     use wiremock::{Mock, ResponseTemplate};
 
     /// 端到端：DELETE .../roles/{role_id} → DeleteResp。
+    /// 同时断言 method、path、auth（catalog 提供）和响应（#438）。
     #[tokio::test]
     async fn test_delete_role_returns_data_on_success() {
         let server = MockServer::start().await;
         Mock::given(method("DELETE"))
             .and(path("/open-apis/base/v2/apps/app001/roles/role001"))
+            .and(header("Authorization", "Bearer test-tenant-token"))
             .respond_with(ResponseTemplate::new(200).set_body_json(json!({
                 "code": 0, "msg": "success", "data": {}
             })))
@@ -102,17 +104,28 @@ mod tests {
             .base_url(server.uri())
             .enable_token_cache(false)
             .build();
+        let option = openlark_core::req_option::RequestOption::builder()
+            .tenant_access_token("test-tenant-token")
+            .build();
         Delete::new(config)
             .app_token("app001")
             .role_id("role001")
-            .execute()
+            .execute_with_options(option)
             .await
             .expect("删除角色应成功");
         let received = server.received_requests().await.unwrap_or_default();
         assert_eq!(received.len(), 1);
+        assert_eq!(received[0].method, "DELETE");
         assert_eq!(
             received[0].url.path(),
             "/open-apis/base/v2/apps/app001/roles/role001"
+        );
+        assert_eq!(
+            received[0]
+                .headers
+                .get("authorization")
+                .and_then(|h| h.to_str().ok()),
+            Some("Bearer test-tenant-token")
         );
     }
 
@@ -121,5 +134,12 @@ mod tests {
         let ep = BaseApiV2::RoleDelete("app".into(), "role001".into());
         let req: openlark_core::api::ApiRequest<DeleteResp> = ep.to_request();
         assert_eq!(req.method(), &openlark_core::api::HttpMethod::Delete);
+        assert_eq!(
+            req.supported_access_token_types(),
+            vec![
+                openlark_core::constants::AccessTokenType::User,
+                openlark_core::constants::AccessTokenType::Tenant
+            ]
+        );
     }
 }
