+++ crates/openlark-docs/src/base/bitable/v1/app/create.rs
@@ -96,11 +96,11 @@ impl CreateAppRequest {
             app_settings: None,
         };
 
-        // 创建API请求 - 使用类型安全的URL生成
-        let api_request: ApiRequest<CreateAppResponse> = ApiRequest::post(&api_endpoint.to_url())
-            .body(openlark_core::api::RequestData::Binary(serde_json::to_vec(
-                &request_body,
-            )?));
+        // #439: method 来自 catalog；叶子不再重复声明稳定请求语义
+        let api_request: ApiRequest<CreateAppResponse> =
+            api_endpoint.to_request::<CreateAppResponse>().body(
+                openlark_core::api::RequestData::Binary(serde_json::to_vec(&request_body)?),
+            );
 
         // 发送请求
         let response = Transport::request(api_request, &self.config, Some(option)).await?;
@@ -142,32 +142,50 @@ mod tests {
     use super::*;
     use serde_json::json;
     use wiremock::MockServer;
-    use wiremock::matchers::{method, path};
+    use wiremock::matchers::{header, method, path};
     use wiremock::{Mock, ResponseTemplate};
 
     /// 端到端：POST /open-apis/bitable/v1/apps → CreateAppResponse。
+    /// 完整断言 method、path、auth（来自 catalog #439）和响应。
     #[tokio::test]
     async fn test_create_app_returns_data_on_success() {
         let server = MockServer::start().await;
         Mock::given(method("POST"))
             .and(path("/open-apis/bitable/v1/apps"))
+            .and(header("Authorization", "Bearer test-tenant-token"))
             .respond_with(ResponseTemplate::new(200).set_body_json(json!({
                 "code": 0, "msg": "success", "data": { "app": { "app_token": "app001", "name": "测试应用" } }
             })))
-            .mount(&server).await;
+            .mount(&server)
+            .await;
         let config = Config::builder()
             .app_id("ci_app_id")
             .app_secret("ci_app_secret")
             .base_url(server.uri())
             .enable_token_cache(false)
             .build();
-        CreateAppRequest::new(config)
+        let option = openlark_core::req_option::RequestOption::builder()
+            .tenant_access_token("test-tenant-token")
+            .build();
+        let resp = CreateAppRequest::new(config)
             .name("测试应用")
-            .execute()
+            .execute_with_options(option)
             .await
             .expect("创建多维表格应成功");
+        assert_eq!(resp.app.app_token, "app001");
         let received = server.received_requests().await.unwrap_or_default();
         assert_eq!(received.len(), 1);
+        assert_eq!(received[0].method, "POST");
         assert_eq!(received[0].url.path(), "/open-apis/bitable/v1/apps");
+        let body: serde_json::Value =
+            serde_json::from_slice(&received[0].body).expect("请求体应为 JSON");
+        assert_eq!(body["name"], "测试应用");
+        assert_eq!(
+            received[0]
+                .headers
+                .get("authorization")
+                .and_then(|h| h.to_str().ok()),
+            Some("Bearer test-tenant-token")
+        );
     }
 }
