+++ crates/openlark-docs/src/base/bitable/v1/app/role/delete.rs
@@ -58,9 +58,9 @@ impl DeleteAppRoleRequest {
 
         use crate::common::api_endpoints::BitableApiV1;
         let api_endpoint = BitableApiV1::RoleDelete(self.app_token, self.role_id);
+        // #439: method 来自 catalog
 
-        let api_request: ApiRequest<DeleteAppRoleResponse> =
-            ApiRequest::delete(&api_endpoint.to_url());
+        let api_request: ApiRequest<DeleteAppRoleResponse> = api_endpoint.to_request();
 
         let response = Transport::request(api_request, &self.config, Some(option)).await?;
         response
