+++ crates/openlark-docs/src/base/bitable/v1/app/role/list.rs
@@ -77,9 +77,9 @@ impl ListAppRoleRequest {
 
         use crate::common::api_endpoints::BitableApiV1;
         let api_endpoint = BitableApiV1::RoleList(self.app_token.clone());
+        // #439: method 来自 catalog
 
-        let mut api_request: ApiRequest<ListAppRoleResponse> =
-            ApiRequest::get(&api_endpoint.to_url());
+        let mut api_request: ApiRequest<ListAppRoleResponse> = api_endpoint.to_request();
         api_request = api_request.query_opt("page_token", self.page_token);
         api_request = api_request.query_opt("page_size", self.page_size.map(|v| v.to_string()));
