+++ crates/openlark-docs/src/base/bitable/v1/app/role/member/batch_create.rs
@@ -90,12 +90,12 @@ impl BatchCreateRoleMemberRequest {
         let api_endpoint =
             BitableApiV1::RoleMemberBatchCreate(self.app_token.clone(), self.role_id.clone());
 
-        let api_request: ApiRequest<BatchCreateRoleMemberResponse> = ApiRequest::post(
-            &api_endpoint.to_url(),
-        )
-        .body(serde_json::to_vec(&BatchCreateRoleMemberRequestBody {
-            member_list: self.member_list,
-        })?);
+        // #439: method 来自 catalog
+        let api_request: ApiRequest<BatchCreateRoleMemberResponse> = api_endpoint
+            .to_request::<BatchCreateRoleMemberResponse>()
+            .body(serde_json::to_vec(&BatchCreateRoleMemberRequestBody {
+                member_list: self.member_list,
+            })?);
 
         let response = Transport::request(api_request, &self.config, Some(option)).await?;
         response
