+++ crates/openlark-docs/src/base/bitable/v1/app/role/member/batch_delete.rs
@@ -90,12 +90,12 @@ impl BatchDeleteRoleMemberRequest {
         let api_endpoint =
             BitableApiV1::RoleMemberBatchDelete(self.app_token.clone(), self.role_id.clone());
 
-        let api_request: ApiRequest<BatchDeleteRoleMemberResponse> = ApiRequest::post(
-            &api_endpoint.to_url(),
-        )
-        .body(serde_json::to_vec(&BatchDeleteRoleMemberRequestBody {
-            member_list: self.member_list,
-        })?);
+        // #439: method 来自 catalog
+        let api_request: ApiRequest<BatchDeleteRoleMemberResponse> = api_endpoint
+            .to_request::<BatchDeleteRoleMemberResponse>()
+            .body(serde_json::to_vec(&BatchDeleteRoleMemberRequestBody {
+                member_list: self.member_list,
+            })?);
 
         let response = Transport::request(api_request, &self.config, Some(option)).await?;
         response
