+++ crates/openlark-docs/src/base/bitable/v1/app/table/batch_create.rs
@@ -87,12 +87,12 @@ impl BatchCreateTableRequest {
         }
 
         let api_endpoint = BitableApiV1::TableBatchCreate(self.app_token);
-        let mut api_request: ApiRequest<BatchCreateTableResponse> = ApiRequest::post(
-            &api_endpoint.to_url(),
-        )
-        .body(serde_json::to_vec(&BatchCreateTableRequestBody {
-            tables: self.tables,
-        })?);
+        // #439: method 来自 catalog
+        let mut api_request: ApiRequest<BatchCreateTableResponse> = api_endpoint
+            .to_request::<BatchCreateTableResponse>()
+            .body(serde_json::to_vec(&BatchCreateTableRequestBody {
+                tables: self.tables,
+            })?);
 
         api_request = api_request.query_opt("user_id_type", self.user_id_type);
         api_request = api_request.query_opt("client_token", self.client_token);
