+++ crates/openlark-docs/src/base/bitable/v1/app/table/batch_delete.rs
@@ -85,12 +85,12 @@ impl BatchDeleteTableRequest {
         }
 
         let api_endpoint = BitableApiV1::TableBatchDelete(self.app_token);
-        let mut api_request: ApiRequest<BatchDeleteTableResponse> = ApiRequest::post(
-            &api_endpoint.to_url(),
-        )
-        .body(serde_json::to_vec(&BatchDeleteTableRequestBody {
-            table_ids: self.table_ids,
-        })?);
+        // #439: method 来自 catalog
+        let mut api_request: ApiRequest<BatchDeleteTableResponse> = api_endpoint
+            .to_request::<BatchDeleteTableResponse>()
+            .body(serde_json::to_vec(&BatchDeleteTableRequestBody {
+                table_ids: self.table_ids,
+            })?);
 
         api_request = api_request.query_opt("user_id_type", self.user_id_type);
         api_request = api_request.query_opt("client_token", self.client_token);
@@ -123,10 +123,11 @@ mod tests {
     use super::*;
     use serde_json::json;
     use wiremock::MockServer;
-    use wiremock::matchers::{method, path};
+    use wiremock::matchers::{header, method, path};
     use wiremock::{Mock, ResponseTemplate};
 
     /// 端到端：POST .../tables/batch_delete → BatchDeleteTableResponse。
+    /// 完整断言 method、path、auth（来自 catalog #439）和响应。
     #[tokio::test]
     async fn test_batch_delete_table_returns_data_on_success() {
         let server = MockServer::start().await;
@@ -134,6 +135,7 @@ mod tests {
             .and(path(
                 "/open-apis/bitable/v1/apps/app001/tables/batch_delete",
             ))
+            .and(header("Authorization", "Bearer test-tenant-token"))
             .respond_with(ResponseTemplate::new(200).set_body_json(json!({
                 "code": 0, "msg": "success", "data": {}
             })))
@@ -145,17 +147,47 @@ mod tests {
             .base_url(server.uri())
             .enable_token_cache(false)
             .build();
-        BatchDeleteTableRequest::new(config)
+        let option = openlark_core::req_option::RequestOption::builder()
+            .tenant_access_token("test-tenant-token")
+            .build();
+        let resp = BatchDeleteTableRequest::new(config)
             .app_token("app001")
+            .user_id_type("open_id")
+            .client_token("client_001")
             .table_ids(vec!["tbl001".into()])
-            .execute()
+            .execute_with_options(option)
             .await
             .expect("批量删除数据表应成功");
+        let _response = resp;
         let received = server.received_requests().await.unwrap_or_default();
         assert_eq!(received.len(), 1);
+        assert_eq!(received[0].method, "POST");
         assert_eq!(
             received[0].url.path(),
             "/open-apis/bitable/v1/apps/app001/tables/batch_delete"
+        );
+        let body: serde_json::Value =
+            serde_json::from_slice(&received[0].body).expect("请求体应为 JSON");
+        assert_eq!(body["table_ids"], json!(["tbl001"]));
+        let query: std::collections::HashMap<_, _> = received[0]
+            .url
+            .query_pairs()
+            .map(|(key, value)| (key.into_owned(), value.into_owned()))
+            .collect();
+        assert_eq!(
+            query.get("user_id_type").map(String::as_str),
+            Some("open_id")
+        );
+        assert_eq!(
+            query.get("client_token").map(String::as_str),
+            Some("client_001")
+        );
+        assert_eq!(
+            received[0]
+                .headers
+                .get("authorization")
+                .and_then(|h| h.to_str().ok()),
+            Some("Bearer test-tenant-token")
         );
     }
 }
