+++ crates/openlark-docs/src/base/bitable/v1/app/table/create.rs
@@ -158,8 +158,9 @@ impl CreateTableRequest {
         let request_body = CreateTableRequestBody { table: self.table };
 
         // 创建API请求 - 使用类型安全的URL生成
-        let api_request: ApiRequest<CreateTableResponse> =
-            ApiRequest::post(&api_endpoint.to_url()).body(serde_json::to_vec(&request_body)?);
+        let api_request: ApiRequest<CreateTableResponse> = api_endpoint
+            .to_request()
+            .body(serde_json::to_vec(&request_body)?);
 
         // 发送请求
         let response = Transport::request(api_request, &self.config, Some(option)).await?;
