+++ crates/openlark-docs/src/base/bitable/v1/app/table/delete.rs
@@ -72,10 +72,8 @@ impl DeleteTableRequest {
         // 替代传统的字符串拼接方式，提供类型安全和IDE自动补全
         use crate::common::api_endpoints::BitableApiV1;
         let api_endpoint = BitableApiV1::TableDelete(self.app_token.clone(), self.table_id.clone());
-
-        // 创建API请求 - 使用类型安全的URL生成
-        let api_request: ApiRequest<DeleteTableResponse> =
-            ApiRequest::delete(&api_endpoint.to_url());
+        // #439: method 来自 catalog
+        let api_request: ApiRequest<DeleteTableResponse> = api_endpoint.to_request();
 
         // 发送请求
         let response = Transport::request(api_request, &self.config, Some(option)).await?;
