+++ crates/openlark-docs/src/base/bitable/v1/app/table/field/delete.rs
@@ -81,8 +81,7 @@ impl DeleteFieldRequest {
         );
 
         // 创建API请求 - 使用类型安全的URL生成
-        let api_request: ApiRequest<DeleteFieldResponse> =
-            ApiRequest::delete(&api_endpoint.to_url());
+        let api_request: ApiRequest<DeleteFieldResponse> = api_endpoint.to_request();
 
         // 发送请求
         let response = Transport::request(api_request, &self.config, Some(option)).await?;
