+++ crates/openlark-docs/src/base/bitable/v1/app/table/form/upgrade.rs
@@ -132,7 +132,7 @@ impl UpgradeFormRequest {
             display_mode,
         };
         let api_request: ApiRequest<UpgradeFormResponse> =
-            ApiRequest::post(&api_endpoint.to_url()).body(serde_json::to_vec(&body)?);
+            api_endpoint.to_request().body(serde_json::to_vec(&body)?);
 
         let response = Transport::request(api_request, &self.config, Some(option)).await?;
         response
