+++ crates/openlark-docs/src/base/bitable/v1/app/table/list.rs
@@ -81,10 +81,8 @@ impl ListTablesRequest {
         // 替代传统的字符串拼接方式，提供类型安全和IDE自动补全
         use crate::common::api_endpoints::BitableApiV1;
         let api_endpoint = BitableApiV1::TableList(self.app_token.clone());
-
-        // 创建API请求 - 使用类型安全的URL生成
-        let mut api_request: ApiRequest<ListTablesResponse> =
-            ApiRequest::get(&api_endpoint.to_url());
+        // #439: method 来自 catalog
+        let mut api_request: ApiRequest<ListTablesResponse> = api_endpoint.to_request();
 
         // 构建查询参数
         if let Some(page_size) = self.page_size {
