+++ crates/openlark-docs/src/base/bitable/v1/app/table/view/create.rs
@@ -133,10 +133,10 @@ impl CreateViewRequest {
         // 替代传统的字符串拼接方式，提供类型安全和IDE自动补全
         use crate::common::api_endpoints::BitableApiV1;
         let api_endpoint = BitableApiV1::ViewCreate(self.app_token.clone(), self.table_id.clone());
-
-        // 创建API请求 - 使用类型安全的URL生成
-        let api_request: ApiRequest<CreateViewResponse> =
-            ApiRequest::post(&api_endpoint.to_url()).body(serde_json::to_vec(&self.view)?);
+        // #439: method 来自 catalog
+        let api_request: ApiRequest<CreateViewResponse> = api_endpoint
+            .to_request()
+            .body(serde_json::to_vec(&self.view)?);
 
         // 发送请求
         let response = Transport::request(api_request, &self.config, Some(option)).await?;
