+++ crates/openlark-docs/src/base/bitable/v1/app/table/view/list.rs
@@ -125,10 +125,8 @@ impl ListViewsRequest {
         // 替代传统的字符串拼接方式，提供类型安全和IDE自动补全
         use crate::common::api_endpoints::BitableApiV1;
         let api_endpoint = BitableApiV1::ViewList(self.app_token.clone(), self.table_id.clone());
-
-        // 创建API请求 - 使用类型安全的URL生成
-        let mut api_request: ApiRequest<ListViewsResponse> =
-            ApiRequest::get(&api_endpoint.to_url());
+        // #439: method 来自 catalog
+        let mut api_request: ApiRequest<ListViewsResponse> = api_endpoint.to_request();
 
         // 构建查询参数
         if let Some(ref user_id_type) = self.user_id_type {
