+++ crates/openlark-docs/src/base/bitable/v1/app/update.rs
@@ -104,11 +104,11 @@ impl UpdateAppRequest {
             app_settings: self.app_settings.clone(),
         };
 
-        // 创建API请求 - 使用类型安全的URL生成
-        let api_request: ApiRequest<UpdateAppResponse> = ApiRequest::put(&api_endpoint.to_url())
-            .body(openlark_core::api::RequestData::Binary(serde_json::to_vec(
-                &request_body,
-            )?));
+        // #439: method 来自 catalog
+        let api_request: ApiRequest<UpdateAppResponse> =
+            api_endpoint.to_request::<UpdateAppResponse>().body(
+                openlark_core::api::RequestData::Binary(serde_json::to_vec(&request_body)?),
+            );
 
         // 发送请求
         let response = Transport::request(api_request, &self.config, Some(option)).await?;
@@ -155,7 +155,7 @@ mod tests {
     use super::*;
     use serde_json::json;
     use wiremock::MockServer;
-    use wiremock::matchers::{method, path};
+    use wiremock::matchers::{header, method, path};
     use wiremock::{Mock, ResponseTemplate};
 
     /// 端到端：PUT .../apps/{app_token} → UpdateAppResponse。
@@ -164,24 +164,41 @@ mod tests {
         let server = MockServer::start().await;
         Mock::given(method("PUT"))
             .and(path("/open-apis/bitable/v1/apps/app001"))
+            .and(header("Authorization", "Bearer test-tenant-token"))
             .respond_with(ResponseTemplate::new(200).set_body_json(json!({
                 "code": 0, "msg": "success", "data": { "app": { "app_token": "app001", "name": "新名称" } }
             })))
-            .mount(&server).await;
+            .mount(&server)
+            .await;
         let config = Config::builder()
             .app_id("ci_app_id")
             .app_secret("ci_app_secret")
             .base_url(server.uri())
             .enable_token_cache(false)
             .build();
-        UpdateAppRequest::new(config)
+        let option = openlark_core::req_option::RequestOption::builder()
+            .tenant_access_token("test-tenant-token")
+            .build();
+        let response = UpdateAppRequest::new(config)
             .app_token("app001")
             .name("新名称")
-            .execute()
+            .execute_with_options(option)
             .await
             .expect("更新多维表格应成功");
+        assert_eq!(response.app.name, "新名称");
         let received = server.received_requests().await.unwrap_or_default();
         assert_eq!(received.len(), 1);
+        assert_eq!(received[0].method, "PUT");
         assert_eq!(received[0].url.path(), "/open-apis/bitable/v1/apps/app001");
+        let body: serde_json::Value =
+            serde_json::from_slice(&received[0].body).expect("请求体应为 JSON");
+        assert_eq!(body["name"], "新名称");
+        assert_eq!(
+            received[0]
+                .headers
+                .get("authorization")
+                .and_then(|value| value.to_str().ok()),
+            Some("Bearer test-tenant-token")
+        );
     }
 }
